2024-01-08 v0.15 Added statusbar icon when auto return to watchface soon

+++ Firmware/DRM_Watch_v3/Button.ino
@@ -37,7 +37,7 @@ void registerAction(){
   lastActionTime=millis();
 }
 
-unsigned long sinceLastAction(){
+int sinceLastAction(){
   return millis() - lastActionTime;
 }
